backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.1
asyncpg==0.30.0
pydantic==2.8.2
pydantic-settings==2.3.4
anthropic==0.29.0
httpx==0.27.0
pandas==2.2.2
openpyxl==3.1.3
phonenumbers==8.13.39
python-multipart==0.0.9
boto3==1.34.144
cryptography==42.0.8
google-auth==2.31.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.139.0
python-dotenv==1.0.1
sse-starlette==2.1.2
